+++ packages/sdk/src/contracts/StreamStorageRegistry.ts
@@ -14,6 +14,7 @@ import { LoggerFactory } from '../utils/LoggerFactory'
 import { ChainEventPoller } from './ChainEventPoller'
 import { ContractFactory } from './ContractFactory'
 import { initContractEventGateway, waitForTx } from './contract'
+import { CachingMap } from '../utils/CachingMap'
 
 export interface StorageNodeAssignmentEvent {
     readonly streamId: StreamID
@@ -26,6 +27,8 @@ interface NodeQueryResult {
     metadata: string
     lastseen: string
 }
+
+const GET_ALL_STORAGE_NODES = Symbol('GET_ALL_STORAGE_NODES')
 
 /**
  * Stores storage node assignments (mapping of streamIds <-> storage nodes addresses)
@@ -42,13 +45,14 @@ export class StreamStorageRegistry {
     private readonly config: Pick<StrictStreamrClientConfig, 'contracts' | '_timeouts'>
     private readonly authentication: Authentication
     private readonly logger: Logger
+    private readonly getStorageNodes_cached: CachingMap<string | typeof GET_ALL_STORAGE_NODES, EthereumAddress[], [string | undefined]>
 
     constructor(
         streamIdBuilder: StreamIDBuilder,
         contractFactory: ContractFactory,
         rpcProviderSource: RpcProviderSource,
         theGraphClient: TheGraphClient,
-        @inject(ConfigInjectionToken) config: Pick<StrictStreamrClientConfig, 'contracts' | '_timeouts'>,
+        @inject(ConfigInjectionToken) config: Pick<StrictStreamrClientConfig, 'contracts' | 'cache' | '_timeouts'>,
         @inject(AuthenticationInjectionToken) authentication: Authentication,
         eventEmitter: StreamrClientEventEmitter,
         loggerFactory: LoggerFactory
@@ -74,6 +78,14 @@ export class StreamStorageRegistry {
             )
         }), config.contracts.pollInterval)
         this.initStreamAssignmentEventListeners(eventEmitter, chainEventPoller, loggerFactory)
+        this.getStorageNodes_cached = new CachingMap((streamIdOrPath?: string) => {
+            return this.getStorageNodes_nonCached(streamIdOrPath)
+        }, {
+            ...config.cache,
+            cacheKey: ([streamIdOrPath]): string | typeof GET_ALL_STORAGE_NODES => {
+                return streamIdOrPath ?? GET_ALL_STORAGE_NODES
+            }
+        })
     }
 
     // eslint-disable-next-line class-methods-use-this
@@ -123,6 +135,7 @@ export class StreamStorageRegistry {
         await this.connectToContract()
         const ethersOverrides = await getEthersOverrides(this.rpcProviderSource, this.config)
         await waitForTx(this.streamStorageRegistryContract!.addStorageNode(streamId, nodeAddress, ethersOverrides))
+        this.getStorageNodes_cached.invalidate((key) => key === streamId)
     }
 
     async removeStreamFromStorageNode(streamIdOrPath: string, nodeAddress: EthereumAddress): Promise<void> {
@@ -131,6 +144,7 @@ export class StreamStorageRegistry {
         await this.connectToContract()
         const ethersOverrides = await getEthersOverrides(this.rpcProviderSource, this.config)
         await waitForTx(this.streamStorageRegistryContract!.removeStorageNode(streamId, nodeAddress, ethersOverrides))
+        this.getStorageNodes_cached.invalidate((key) => key === streamId)
     }
 
     async isStoredStream(streamIdOrPath: string, nodeAddress: EthereumAddress): Promise<boolean> {
@@ -178,6 +192,10 @@ export class StreamStorageRegistry {
     }
 
     async getStorageNodes(streamIdOrPath?: string): Promise<EthereumAddress[]> {
+        return this.getStorageNodes_cached.get(streamIdOrPath)
+    }
+
+    private async getStorageNodes_nonCached(streamIdOrPath?: string): Promise<EthereumAddress[]> {
         let queryResults: NodeQueryResult[]
         if (streamIdOrPath !== undefined) {
             const streamId = await this.streamIdBuilder.toStreamID(streamIdOrPath)