+++ packages/sdk/test/unit/Operator.test.ts
@@ -14,23 +14,23 @@ import { mockLoggerFactory } from '../test-utils/utils'
 
 describe(parsePartitionFromReviewRequestMetadata, () => {
     it('throws given undefined', () => {
-        expect(() => parsePartitionFromReviewRequestMetadata(undefined)).toThrowError(ParseError)
+        expect(() => parsePartitionFromReviewRequestMetadata(undefined)).toThrow(ParseError)
     })
 
     it('throws given invalid json', () => {
-        expect(() => parsePartitionFromReviewRequestMetadata('invalidjson')).toThrowError(ParseError)
+        expect(() => parsePartitionFromReviewRequestMetadata('invalidjson')).toThrow(ParseError)
     })
 
     it('throws given valid json without field "partition"', () => {
-        expect(() => parsePartitionFromReviewRequestMetadata('{}')).toThrowError(ParseError)
+        expect(() => parsePartitionFromReviewRequestMetadata('{}')).toThrow(ParseError)
     })
 
     it('throws given valid json with field "partition" but not as a number', () => {
-        expect(() => parsePartitionFromReviewRequestMetadata('{ "partition": "foo" }')).toThrowError(ParseError)
+        expect(() => parsePartitionFromReviewRequestMetadata('{ "partition": "foo" }')).toThrow(ParseError)
     })
 
     it('throws given valid json with field "partition" but outside integer range', () => {
-        expect(() => parsePartitionFromReviewRequestMetadata('{ "partition": -50 }')).toThrowError(ParseError)
+        expect(() => parsePartitionFromReviewRequestMetadata('{ "partition": -50 }')).toThrow(ParseError)
     })
 
     it('returns partition given valid json with field "partition" within integer range', () => {