+++ packages/sdk/test/unit/messagePipeline.test.ts
@@ -9,7 +9,7 @@ import { StrictStreamrClientConfig } from '../../src/Config'
 import { DestroySignal } from '../../src/DestroySignal'
 import { ERC1271ContractFacade } from '../../src/contracts/ERC1271ContractFacade'
 import { StreamRegistry } from '../../src/contracts/StreamRegistry'
-import { DecryptError, EncryptionUtil } from '../../src/encryption/EncryptionUtil'
+import { EncryptionUtil } from '../../src/encryption/EncryptionUtil'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { GroupKeyManager } from '../../src/encryption/GroupKeyManager'
 import { LitProtocolFacade } from '../../src/encryption/LitProtocolFacade'
@@ -22,6 +22,7 @@ import { PushPipeline } from '../../src/utils/PushPipeline'
 import { mockLoggerFactory } from '../test-utils/utils'
 import { MessageID } from './../../src/protocol/MessageID'
 import { ContentType, EncryptionType, SignatureType, StreamMessage, StreamMessageType } from './../../src/protocol/StreamMessage'
+import { StreamrClientError } from '../../src/StreamrClientError'
 
 const CONTENT = {
     foo: 'bar'
@@ -77,7 +78,7 @@ describe('messagePipeline', () => {
         streamRegistry = {
             getStreamMetadata: async () => ({ partitions: 1 }),
             isStreamPublisher: async () => true,
-            invalidateStreamCache: jest.fn()
+            invalidatePermissionCaches: jest.fn()
         }
         pipeline = createMessagePipeline({
             streamPartId,
@@ -133,7 +134,7 @@ describe('messagePipeline', () => {
         const onError = jest.fn()
         pipeline.onError.listen(onError)
         const output = await collect(pipeline)
-        expect(onError).toBeCalledTimes(1)
+        expect(onError).toHaveBeenCalledTimes(1)
         const error = onError.mock.calls[0][0]
         expect(error.message).toContain('Signature validation failed')
         expect(output).toEqual([])
@@ -148,7 +149,7 @@ describe('messagePipeline', () => {
         const onError = jest.fn()
         pipeline.onError.listen(onError)
         const output = await collect(pipeline)
-        expect(onError).toBeCalledTimes(1)
+        expect(onError).toHaveBeenCalledTimes(1)
         const error = onError.mock.calls[0][0]
         expect(error.message).toContain('Invalid JSON')
         expect(output).toEqual([])
@@ -166,13 +167,14 @@ describe('messagePipeline', () => {
         const onError = jest.fn()
         pipeline.onError.listen(onError)
         const output = await collect(pipeline)
-        expect(onError).toBeCalledTimes(1)
+        expect(onError).toHaveBeenCalledTimes(1)
         const error = onError.mock.calls[0][0]
-        expect(error).toBeInstanceOf(DecryptError)
-        expect(error.message).toMatch(/timed out/)
+        expect(error).toBeInstanceOf(StreamrClientError)
+        expect(error.code).toBe('DECRYPT_ERROR')
+        expect(error.message).toMatch(/Could not get encryption key/)
         expect(output).toEqual([])
-        expect(streamRegistry.invalidateStreamCache).toBeCalledTimes(1)
-        expect(streamRegistry.invalidateStreamCache).toBeCalledWith(StreamPartIDUtils.getStreamID(streamPartId))
+        expect(streamRegistry.invalidatePermissionCaches).toHaveBeenCalledTimes(1)
+        expect(streamRegistry.invalidatePermissionCaches).toHaveBeenCalledWith(StreamPartIDUtils.getStreamID(streamPartId))
     })
 
     it('error: exception', async () => {
@@ -184,7 +186,7 @@ describe('messagePipeline', () => {
         pipeline.onError.listen(onError)
         const output = await collect(pipeline)
         expect(output).toHaveLength(1)
-        expect(onError).toBeCalledTimes(1)
-        expect(onError).toBeCalledWith(err)
+        expect(onError).toHaveBeenCalledTimes(1)
+        expect(onError).toHaveBeenCalledWith(err)
     })
 })